+++ filter.js
@@ -1,6 +1,6 @@
 /*------------------------FILTER VALUES------------------------*/
 
-let filter = {
+export let filter = {
     online: true, //online challenges checkbox
     onsite: true, //onsite challenges checkbox
     minRating: 1, //lowest rating
@@ -9,7 +9,7 @@ let filter = {
     words: [], //inputfield words split into strings in lowercase
   };
 
-  export default filter;
+  
   
   /*------------------------FILTER VALUES------------------------*/
   
@@ -103,102 +103,10 @@ let filter = {
   
   /*------------------------TAGS------------------------*/
   
-  // Get all the tags
-    
-
-
-  const tags = [
-    document.querySelector("#web"),
-    document.querySelector("#linux"),
-    document.querySelector("#cryptography"),
-    document.querySelector("#coding"),
-    document.querySelector("#ssh"),
-    document.querySelector("#ctf"),
-    document.querySelector("#hacking"),
-    document.querySelector("#bash"),
-    document.querySelector("#javascript"),
-    document.querySelector("#electronics"),
-    document.querySelector("#phreaking"),
-  ];
   
-  // Add event listeners to all tags
-  tags.forEach((tag) => {
-    tag.addEventListener("click", () => handleTagClick(tag));
-  });
-  
-  function handleTagClick(tag) {
-    // Get the id of the clicked tag
-    const tagId = tag.id;
-  
-    // Check if the clicked tag has the 'active' class
-    if (tag.classList.contains("active")) {
-      // If it does, remove the 'active' class
-      tag.classList.remove("active");
-      // And remove the tag's id from the filter.labels array
-      filter.labels = filter.labels.filter((activeTag) => activeTag !== tagId);
-      // Log a message indicating that the tag was removed
-      console.log("Removed tag:", tagId);
-    } else {
-      // If the clicked tag does not have the 'active' class, add it
-      tag.classList.add("active");
-      // And add the tag's id to the filter.labels array
-      filter.labels.push(tagId);
-      
-    }
-  
-    //Log the current state of the filter.labels array
-    //console.log('Current labels:', filter.labels);
-  
-        
-    //Get all the challenge elements
-    const card = document.querySelectorAll(".challenge");
-
-    //Get all the tags within the challenge elements
-    const challengeTags = document.querySelectorAll("challenge.tags");
-    
-
-    //Get the noMatchError element
-    const noMatchError = document.getElementById("noMatchError");
-
-    // Initialize a variable to track whether a match is found
-    let foundMatch = false;
-
-    // saves the active labels array into a variable
-    const activeTags = filter.labels;
-   
-
-    // Loop through all the challenge elements
-    for (let i = 0; i < card.length; i++) {
-      // Get the tag in the current challenge element
-      let cardTag = challengeTags[i].tags;
-
-      //If the tag inside the challenge card includes the active tag value
-      if (cardTag.toLowerCase().indexOf(activeTags) > -1 ) {
-        // Show the current challenge element
-        card[i].style.display = "";
-
-        // Set foundMatch to true since a match was found
-        foundMatch = true;
-      } else {
-        // Hide the current challenge element if it doesn't match the search box value
-        card[i].style.display = "none";
-      }
-      // If a match was found
-      if (foundMatch) {
-        // Clear the error message
-        noMatchError.innerHTML = "";
-      } else {
-        // If no matches were found, set the error message
-        noMatchError.innerHTML = "No matching challenges";
-      }
-    }
-  }
-
-  //Endast rum med aktiva tags ska visas - finns i array filter.labels
- //challenge.tags (tagsen som är på korten)
  
-
  
+  
   /*------------------------TAGS------------------------*/
   
   /*------------------------INPUT FIELD------------------------*/
@@ -278,5 +186,3 @@ function searchRooms() {
     noMatchError.innerHTML = "No matching challenges";
   }
 }
-
-  /*------------------------INPUT FIELD------------------------*/
